Implementando Arquitetura REST e Serializers

+++ projeto/monitoramento/models.py
@@ -3,7 +3,7 @@ from django.db import models
 class Leituras(models.Model):
     temperatura = models.FloatField()
     ph = models.FloatField()
-    tds = models.FloatField()
+    tds = models.IntegerField()
     data_hora = models.DateTimeField(auto_now_add=True)
 
 def __str__(self):
